+++ python/GafferDeadlineTest/DeadlineDispatcherTest.py
@@ -36,11 +36,7 @@
 
 import os
 import unittest
-import sys
-if sys.version_info >= (3, 3):
-    from unittest import mock
-else:
-    import mock
+from unittest import mock
 
 import IECore